Move types to config.go
